+++ tensorflow/core/kernels/gather_nd_op.cc
@@ -16,13 +16,11 @@ limitations under the License.
 // See docs in ../ops/array_ops.cc.
 #define EIGEN_USE_THREADS
 
-#include <atomic>
-
+#include "tensorflow/core/kernels/gather_nd_op.h"
 #include "tensorflow/core/framework/op_kernel.h"
 #include "tensorflow/core/framework/register_types.h"
 #include "tensorflow/core/framework/tensor.h"
 #include "tensorflow/core/kernels/bounds_check.h"
-#include "tensorflow/core/kernels/gather_nd_op.h"
 #include "tensorflow/core/platform/logging.h"
 #include "tensorflow/core/platform/mem.h"
 #include "tensorflow/core/platform/types.h"
@@ -154,97 +152,6 @@ class GatherNdOp : public OpKernel {
     }
   }
 };
-
-// Specialization of GatherNdSlice to CPU
-namespace generator {
-
-template <typename T, typename Index, int IXDIM>
-class GatherNdSliceGenerator {
- public:
-  EIGEN_DEVICE_FUNC EIGEN_ALWAYS_INLINE GatherNdSliceGenerator(
-      const Index slice_size, typename TTypes<Index>::ConstMatrix Tindices,
-      typename TTypes<T, IXDIM + 1>::ConstTensor Tparams,
-      typename TTypes<T>::Matrix Tout, std::atomic<Index>* error_loc)
-      : slice_size_(slice_size),
-        Tindices_(Tindices),
-        Tparams_(Tparams),
-        Tout_(Tout),
-        error_loc_(error_loc) {}
-
-  EIGEN_DEVICE_FUNC EIGEN_ALWAYS_INLINE bool GenerateIndices(
-      const Index loc, Eigen::array<Eigen::DenseIndex, IXDIM + 1>* ix) const {
-    (*ix)[IXDIM] = 0;
-    bool out_of_bounds = false;
-    for (int i = 0; i < IXDIM; ++i) {
-      const Index ix_i = internal::SubtleMustCopy(Tindices_(loc, i));
-      (*ix)[i] = ix_i;
-      out_of_bounds |= !FastBoundsCheck(ix_i, Tparams_.dimension(i));
-    }
-    return out_of_bounds;
-  }
-
-  EIGEN_DEVICE_FUNC EIGEN_ALWAYS_INLINE int32
-  operator()(const Eigen::array<Eigen::DenseIndex, 1>& loc_array) const {
-    const Index loc = loc_array[0];
-    Eigen::array<Eigen::DenseIndex, IXDIM + 1> ix;
-    Eigen::array<Eigen::DenseIndex, 2> ix_out;
-    ix_out[0] = loc;
-    ix_out[1] = 0;
-    const bool out_of_bounds = GenerateIndices(loc, &ix);
-    if (TF_PREDICT_FALSE(out_of_bounds)) {
-      error_loc_->store(loc);
-      std::fill_n(&Tout_(ix_out), slice_size_, T());
-    } else {
-      std::copy_n(&Tparams_(ix), slice_size_, &Tout_(ix_out));
-    }
-
-    return static_cast<int32>(0);  // Return something...
-  }
-
- private:
-  const Index slice_size_;
-  const typename TTypes<Index>::ConstMatrix Tindices_;
-  const typename TTypes<T, IXDIM + 1>::ConstTensor Tparams_;
-  mutable typename TTypes<T>::Matrix Tout_;
-  std::atomic<Index>* error_loc_;
-};
-
-}  // namespace generator
-
-namespace functor {
-
-template <typename T, typename Index, int IXDIM>
-struct GatherNdSlice<CPUDevice, T, Index, IXDIM> {
-  Index operator()(const CPUDevice& d, const Index slice_size,
-                   typename TTypes<int32>::Scalar Tscratch,
-                   typename TTypes<T, IXDIM + 1>::ConstTensor Tparams,
-                   typename TTypes<Index>::ConstMatrix Tindices,
-                   typename TTypes<T>::Matrix Tout) {
-    std::atomic<Index> error_loc(-1);
-
-    const Eigen::DenseIndex batch_size = Tindices.dimension(0);
-#if !defined(EIGEN_HAS_INDEX_LIST)
-    Eigen::Tensor<Eigen::DenseIndex, 1>::Dimensions reshape_dims{{ 1 }};
-    Eigen::array<Eigen::DenseIndex, 1> broadcast_dims{{ batch_size }};
-#else
-    Eigen::IndexList<Eigen::type2index<1> > reshape_dims;
-    Eigen::IndexList<Eigen::DenseIndex> broadcast_dims;
-    broadcast_dims.set(0, batch_size);
-#endif
-    generator::GatherNdSliceGenerator<T, Index, IXDIM> gather_nd_generator(
-        slice_size, Tindices, Tparams, Tout, &error_loc);
-    Tscratch.device(d) = Tscratch.reshape(reshape_dims)
-                             .broadcast(broadcast_dims)
-                             .generate(gather_nd_generator)
-                             .sum();
-
-    // error_loc() returns -1 if there's no out-of-bounds index,
-    // otherwise it returns the location of an OOB index in Tindices.
-    return error_loc.load();
-  }
-};
-
-}  // namespace functor
 
 #define REGISTER_GATHER_ND_FULL(dev, type, index_type)                 \
   REGISTER_KERNEL_BUILDER(Name("GatherNd")                             \
